mcp_tools/notion/pyproject.toml
[project]
name = "notion-api-mcp"
version = "0.1.0"
description = "Enhanced Notion MCP integration with advanced features"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",  # MCP SDK with CLI features
    "httpx",            # Async HTTP client
    "python-dotenv",    # Environment management
    "pydantic",         # Data validation
    "pytest",           # Testing
    "pytest-asyncio",   # Async test support
    "pytest-cov",       # Test coverage
    "rich",            # Enhanced terminal output
    "structlog"        # Structured logging
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/notion_api_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=notion_api_mcp --cov-report=term-missing"
markers = [
    "integration: marks tests that require Notion API access",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.black]
line-length = 88
target-version = ['py310']